Implementing lb, lw, sb, and sw.

+++ TACSimulator/include/Simulator/Simulator.h
@@ -36,7 +36,7 @@ public:
 private:
 
     std::unique_ptr<std::array<uint8_t, 0xFFFF>> RAM;
-    std::array<int16_t, 32> registers;
+    std::array<int32_t, 32> registers;
     const uint8_t IPRegister = 31;
     const uint8_t TrapRegister = 27;
     const uint8_t RETRegister = 30;
